refactor: 대시보드 목록 조회 로직을 Sidebar에서 AppShell로 이동
