+++ src/components/layout/Sidebar/index.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { useState, useEffect } from 'react';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 import LogoIcon from '@/public/images/icon-logo.svg';
@@ -8,35 +7,9 @@ import PlusIcon from '@/src/components/icons/icon-plus.svg';
 import CrownIcon from '@/src/components/icons/icon-crown.svg';
 import * as S from './styles';
 import type { SidebarProps } from '@/src/components/layout/Sidebar/type';
-import type { Dashboard } from '@/src/apis/dashboards/type';
-import { getDashboardList } from '@/src/apis/dashboards';
 
-export default function Sidebar({ isOpen, onClose }: SidebarProps) {
+export default function Sidebar({ isOpen, onClose, dashboards, isLoading, isError }: SidebarProps) {
   const pathname = usePathname();
-
-  const [dashboards, setDashboards] = useState<Dashboard[]>([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [isError, setIsError] = useState(false);
-
-  useEffect(() => {
-    const fetchDashboards = async () => {
-      setIsLoading(true);
-      setIsError(false);
-
-      try {
-        const { dashboards } = await getDashboardList({ size: 20 }); //TODO 추후 무한 스크롤 구현을 위한 임의의 size 설정
-        setDashboards(dashboards);
-      } catch (e) {
-        console.error(e);
-        setIsError(true);
-        setDashboards([]);
-      } finally {
-        setIsLoading(false);
-      }
-    };
-
-    fetchDashboards();
-  }, []);
 
   return (
     <S.Wrapper $isOpen={isOpen}>
