+++ src/components/layout/Sidebar/type.ts
@@ -1,4 +1,9 @@
+import type { Dashboard } from '@/src/apis/dashboards/type';
+
 export interface SidebarProps {
   isOpen: boolean;
   onClose: () => void;
+  dashboards: Dashboard[];
+  isLoading: boolean;
+  isError: boolean;
 }
